requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pytesseract==0.3.13
Pillow==11.1.0
playwright==1.49.1
httpx==0.28.1
anthropic==0.43.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
